First commit -> re-architect the python repo to be (more) a python project (setup.[py|cfg], Makefile, MANIFEST.in, requirements, ...).
