+++ setup.cfg
@@ -1,0 +1,30 @@
+[bdist_wheel]
+universal=1
+
+[flake8]
+max-line-length = 160
+
+[tool:pytest]
+testpaths = tests
+
+python_files =
+    test_*.py
+    *_test.py
+    tests.py
+
+addopts =
+    -ra
+    --strict
+    --doctest-modules
+    --doctest-glob=\*.rst
+    --tb=short
+
+
+[check-manifest]
+ignore =
+  .dockerignore
+  .gitmodules
+  MANIFEST.in
+  Makefile
+  docker
+  docker/Dockerfile
