+++ tox.ini
@@ -1,0 +1,27 @@
+[tox]
+envlist =
+    check,
+    py3,
+
+[testenv]
+install_command = pip install \
+                        {opts} {packages}
+deps =
+    pytest
+	tox
+
+commands =
+    pytest {posargs}
+
+[testenv:check]
+deps =
+    docutils
+    check-manifest
+    flake8
+    # isort
+skip_install = true
+commands =
+    -python setup.py check --strict --metadata --restructuredtext
+    -check-manifest {toxinidir}
+    -flake8 src tests setup.py
+    # -isort --verbose --check-only --diff --recursive src tests setup.py
